emacs: rewrite the whole dang config
Ok look. I don't think I'm going to become an emacs user again. But I've
been writing a lot of lisp lately, and doing so in emacs is just way
nicer than doing so in vim, so I might as well have a config that's not
totally awful. Here we get rid of all the weird org-babel stuff and
update to a more modern use-package approach. (The old settings.org file
is still here, in case I want it later, but everything that still seemed
load-bearing I've moved into init.el and mjm/whatever packages.)