vim: fix listchars for go
I had set this up (correctly) to show spaces for tabs, but it was being
blatted as soon as I went to insert mode and back by my whitespace
thing, which set the global listchars, overwriting the local ones. Easy
enough to fix, once I figured out what was wrong!