vim: restore save/restore winview on buffer switch
This reverts commit 2e3378e0 (vim: remove some ancient copypasta,
2025-03-08). Turns out, this _was_ useful and I just didn't understand
it! It prevents a goofy situation like this:

Say your cursor is on the last line of a file that's more than one
screen long, and you hit `zz` to bring that line to the middle of the
screen.  If you switch buffers away and then back to that file, by
default vim repositions that line on the last line of the window again.
That's really obnoxious, because it means things jump around when you're
switching buffers a lot.

This thing, which _was_ ancient copypasta, prevents that problem, but I
didn't really understand what it did until months later because I had
been low-key annoyed at this problem for months!